Put events writer and reader into io package.
