+++ src/io.rs
@@ -2,6 +2,7 @@ pub mod matsim_id;
 pub mod network;
 pub mod non_blocking_io;
 pub mod population;
+mod proto_events;
 mod worker;
 mod xml_reader;
 mod xml_writer;
