+++ src/mpi.rs
@@ -1,7 +1,7 @@
 pub mod controller;
+pub mod events;
 mod message_broker;
 pub mod messages;
 mod population;
 mod simulation;
 mod time_queue;
-mod events;
